edit set_up.sh and change model option to cat5

+++ app/main.py
@@ -38,7 +38,7 @@ def upload_file():
         semantic_segment('upload/content/in1.' + str(content.filename).split(".")[-1], 'content')
         semantic_segment('upload/style/in1.' + str(style.filename).split(".")[-1], 'style')
 
-        output[0] = "in1_sum_encoder_skip." + str(content.filename).split(".")[-1]
+        output[0] = "in1." + str(content.filename).split(".")[-1]
 
         return redirect(url_for('stylize', output = output))
 
@@ -51,23 +51,24 @@ def semantic_segment(originPath, resultPath):
 
 @app.route('/stylization')
 def stylize():
-      #hard code
-      os.system('python transfer.py --option_unpool sum -e -s --content ./upload/content --style ./upload/style --content_segment ./segment_res/content --style_segment ./segment_res/style/ --output ./static/ --verbose --image_size 512')
-      res_path = 'static/' + output[0]
-      return render_template("result.html", source = res_path)
+    #hard code
+	os.system('python transfer.py --option_unpool cat5 -e -s --content ./upload/content --style ./upload/style --content_segment ./segment_res/content --style_segment ./segment_res/style/ --output ./static/ --verbose --image_size 512')
+	res_path = 'static/' + output[0]
+	return render_template("result.html", source = res_path)
 
 def clean_up():
-    folders = ['./upload/content/', './upload/style/', 'static/', 'segment_res/content/', 'segment_res/style/']
-    for folder in folders:
-        for filename in os.listdir(folder):
-            file_path = os.path.join(folder, filename)
-            try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as e:
-                print('Failed to delete %s. Reason: %s' % (file_path, e))
+	folders = ['./upload/content/', './upload/style/', 'static/', 'segment_res/content/', 'segment_res/style/']
+	for folder in folders:
+		if not os.path.isdir(folder):
+			for filename in os.listdir(folder):
+			    file_path = os.path.join(folder, filename)
+			    try:
+			        if os.path.isfile(file_path) or os.path.islink(file_path):
+			            os.unlink(file_path)
+			        elif os.path.isdir(file_path):
+			            shutil.rmtree(file_path)
+			    except Exception as e:
+			        print('Failed to delete %s. Reason: %s' % (file_path, e))
 
 if __name__ == '__main__':
     clean_up() ##delet all files from previous session
